move tooltip card to tooltip folder
