+++ src/locales/es/messages.js
@@ -173,6 +173,7 @@
       "<1/>",
       ["2"],
     ],
+    "Current Weight": "Current Weight",
     Dashboard: "Panel",
     Dashboards: "Paneles",
     Date: "Fecha",
@@ -336,6 +337,7 @@
     "MetaMask not detected.": "MetaMask no detectado.",
     "Min leverage: 1.1x": "Apalancamiento Mín.: 1.1x",
     "Min order: 10 USD": "Orden Mín. 10 USD",
+    "More Info": "More Info",
     "Multiplier Points APR": "APR de Puntos Multiplicadores",
     Multiswap: "Multi-intercambio",
     "Net Value": "Valor Neto",
@@ -514,6 +516,7 @@
     "Switch Network": "Cambiar Red",
     Synapse: "Synapse",
     TOKEN: "TOKEN",
+    "Target Weight": "Target Weight",
     "Telegram Groups": "Grupos de Telegram",
     "Terms and Conditions": "Términos y Condiciones",
     "The Boosted APR is from your staked Multiplier Points.":
@@ -723,6 +726,21 @@
       ["0"],
       " GMX tokens can be claimed, use the options under the Total Rewards section to claim them.",
     ],
+    "{0} is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for {1}.": [
+      ["0"],
+      " is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for ",
+      ["1"],
+      ".",
+    ],
+    "{0} is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with {1},  and to <3>swap</3> {2} for other tokens.":
+      [
+        ["0"],
+        " is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with ",
+        ["1"],
+        ",  and to <3>swap</3> ",
+        ["2"],
+        " for other tokens.",
+      ],
     "{0} pool exceeded, try different token": [["0"], " reserva superada, intente otro token"],
     "{0} selected in order form": [["0"], " seleccionado en el formulario de órden"],
     "{0} {1} not supported": [["0"], " ", ["1"], " no admitidos"],
