+++ src/locales/ja/messages.js
@@ -170,6 +170,7 @@
       "<1/>",
       ["2"],
     ],
+    "Current Weight": "Current Weight",
     Dashboard: "ダッシュボード",
     Dashboards: "ダッシュボード",
     Date: "日付",
@@ -325,6 +326,7 @@
     "MetaMask not detected.": "メタマスクが検出できません。",
     "Min leverage: 1.1x": "最小レバレッジ: 1.1倍",
     "Min order: 10 USD": "最小オーダー金額: 10 USD",
+    "More Info": "More Info",
     "Multiplier Points APR": "マルチプライヤーポイントのAPR(実質年利)",
     Multiswap: "マルチスワップ",
     "Net Value": "ネット金額",
@@ -481,6 +483,7 @@
     "Switch Network": "ネットワーク変更",
     Synapse: "Synapse",
     TOKEN: "トークン",
+    "Target Weight": "Target Weight",
     "Telegram Groups": "Telegramグループ",
     "Terms and Conditions": "規約",
     "The Boosted APR is from your staked Multiplier Points.":
@@ -678,6 +681,21 @@
       ["0"],
       " GMX tokens can be claimed, use the options under the Total Rewards section to claim them.",
     ],
+    "{0} is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for {1}.": [
+      ["0"],
+      " is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for ",
+      ["1"],
+      ".",
+    ],
+    "{0} is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with {1},  and to <3>swap</3> {2} for other tokens.":
+      [
+        ["0"],
+        " is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with ",
+        ["1"],
+        ",  and to <3>swap</3> ",
+        ["2"],
+        " for other tokens.",
+      ],
     "{0} pool exceeded, try different token": [["0"], "プールは超過しています。別のトークンを試してください。"],
     "{0} selected in order form": [["0"], "がオーダーフォームで選択されました"],
     "{0} {1} not supported": [["0"], " ", ["1"], "はサポートされていません"],
