+++ src/locales/ko/messages.js
@@ -179,6 +179,7 @@
       "><1/>",
       ["2"],
     ],
+    "Current Weight": "Current Weight",
     Dashboard: "대시보드",
     Dashboards: "대시보드",
     Date: "날짜",
@@ -335,6 +336,7 @@
     "MetaMask not detected.": "메타마스크가 탐지되지 않았습니다.",
     "Min leverage: 1.1x": "최소 레버리지: 1.1x",
     "Min order: 10 USD": "최소 주문: 10 USD",
+    "More Info": "More Info",
     "Multiplier Points APR": "멀티플라이어 포인트 APR",
     Multiswap: "멀티스왑",
     "Net Value": "순이익",
@@ -483,6 +485,7 @@
     "Switch Network": "네트워크 변경",
     Synapse: "Synapse",
     TOKEN: "토큰",
+    "Target Weight": "Target Weight",
     "Telegram Groups": "텔레그램 그룹",
     "Terms and Conditions": "약관",
     "The Boosted APR is from your staked Multiplier Points.":
@@ -675,6 +678,21 @@
       ["0"],
       " GMX tokens can be claimed, use the options under the Total Rewards section to claim them.",
     ],
+    "{0} is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for {1}.": [
+      ["0"],
+      " is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for ",
+      ["1"],
+      ".",
+    ],
+    "{0} is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with {1},  and to <3>swap</3> {2} for other tokens.":
+      [
+        ["0"],
+        " is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with ",
+        ["1"],
+        ",  and to <3>swap</3> ",
+        ["2"],
+        " for other tokens.",
+      ],
     "{0} pool exceeded, try different token": [["0"], " 풀이 초과되었습니다. 다른 자산으로 시도해보세요"],
     "{0} selected in order form": [["0"], "이(가) 주문 폼에서 선택되었습니다."],
     "{0} {1} not supported": [["0"], " ", ["1"], "은(는) 지원되지 않습니다"],
