+++ src/locales/ru/messages.js
@@ -172,6 +172,7 @@
       "<1/>",
       ["2"],
     ],
+    "Current Weight": "Current Weight",
     Dashboard: "Приборная доска",
     Dashboards: "Dashboards",
     Date: "Date",
@@ -329,6 +330,7 @@
     "MetaMask not detected.": "MetaMask not detected.",
     "Min leverage: 1.1x": "Min leverage: 1.1x",
     "Min order: 10 USD": "Min order: 10 USD",
+    "More Info": "More Info",
     "Multiplier Points APR": "Multiplier Points APR",
     Multiswap: "Multiswap",
     "Net Value": "Net Value",
@@ -494,6 +496,7 @@
     "Switch Network": "Switch Network",
     Synapse: "Synapse",
     TOKEN: "TOKEN",
+    "Target Weight": "Target Weight",
     "Telegram Groups": "Telegram Groups",
     "Terms and Conditions": "Terms and Conditions",
     "The Boosted APR is from your staked Multiplier Points.": "The Boosted APR is from your staked Multiplier Points.",
@@ -691,6 +694,21 @@
       ["0"],
       " GMX tokens can be claimed, use the options under the Total Rewards section to claim them.",
     ],
+    "{0} is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for {1}.": [
+      ["0"],
+      " is above its target weight.<0/><1/>Get lower fees to <2>swap</2> tokens for ",
+      ["1"],
+      ".",
+    ],
+    "{0} is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with {1},  and to <3>swap</3> {2} for other tokens.":
+      [
+        ["0"],
+        " is below its target weight.<0/><1/>Get lower fees to <2>buy GLP</2> with ",
+        ["1"],
+        ",  and to <3>swap</3> ",
+        ["2"],
+        " for other tokens.",
+      ],
     "{0} pool exceeded, try different token": [["0"], " pool exceeded, try different token"],
     "{0} selected in order form": [["0"], " selected in order form"],
     "{0} {1} not supported": [["0"], " ", ["1"], " not supported"],
